trying to rename away from event_services_types in favor of `marqo_service_types`
